Corrected the fit_prms noise type from int to float
The variable type of the fit_prms.noise field was incorrect. I changed it from int to float. This was a problem in the case where it was used as the ACF cutoff pwr and if it was 0 due to truncation.

+++ codebase/superdarn/src.lib/tk/fitacf_v3.0/src/fit_structures.h
@@ -86,7 +86,7 @@ typedef struct fit_prms {
   int cp;
   int xcf;
   int tfreq;
-  int noise;
+  float noise;
   int nrang;
   int smsep;
   int nave;
